backend/requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
drf-yasg==1.21.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
Pillow==9.4.0
